Use semver to parse the log version in testing
Instead of comparing the project version and the version read from the
container log directly as strings it makes sense to parse them and
contextually compare them. This is important if you are comparing, for
example, a Python package version that should match the Docker image,
but the Docker image has a build identifier (ex: `1.0.0` and
`1.0.0+build.1`). The version should be considered equivalent in this
scenario if a project is properly following SemVer versioning.