Remove Pipfile checking
Pipenv has deprecated the `pipenv check` command and un-vendored the
safety package that is used to check Python dependencies. Additionally
the safety package itself has moved to a model that requires the use of
an API key to function. With these changes in mind we remove the use of
`pipenv check` from the Dockerfile configuration.