Break the build and push step into separate steps
There is an issue where an image build taking longer than 10 minutes runs
afoul of the GitHub Container Registry token duration. We are breaking
this step up into a build step that does not push and then a push step
that pushes the built images that are locally cached to hopefully work
around this problem.