Adjust the labels dependabot uses
Instead of using the default labels for the `github-actions` package
ecosystem we specify the labels explicitly. This is done to ensure that
dependabot uses our `github-actions` label instead of the default of
`github_actions`. We must also explicitly specify the `dependencies`
label since we are overriding the default label values.